requirements.txt
PyPDF2==3.0.1
pdfplumber==0.10.3
nltk==3.8.1
spacy==3.7.2
scikit-learn==1.3.2
pandas==2.1.3
streamlit>=1.29.0
joblib==1.3.2
groq>=0.9.0
python-dotenv==1.0.0
faiss-cpu==1.8.0
sentence-transformers==2.5.1
fpdf2>=2.7.9
